requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
cycler==0.10.0
discord==1.7.3
discord.py==1.7.3
future==0.18.2
idna==3.2
kiwisolver==1.3.2
matplotlib==3.4.3
MouseInfo==0.1.3
multidict==5.1.0
nest-asyncio==1.5.1
pefile==2021.9.3
Pillow==8.3.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pywin32-ctypes==0.2.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
